use singular file for rust int tests, keeps state
