+++ package.json
@@ -16,7 +16,7 @@
     "test:unit": "yarn build:contract:debug && asp && jest",
     "test:integration": "yarn build:contract:debug && npm run test:integration:ts && npm run test:integration:rs",
     "test:integration:ts": "ava --verbose",
-    "test:integration:rs": "cd src/tests/integration-tests/rs && cargo run --example message && cargo run --example messages"
+    "test:integration:rs": "cd src/tests/integration-tests/rs && cargo run --example integration-tests"
   },
   "devDependencies": {
     "@babel/core": "^7.17.0",
